refactor: replace authentication-repository to authentication-by-account-repository
