+++ src/__tests__/use-case/session/session-by-account/db-session-by-account.spec.ts
@@ -1,4 +1,4 @@
-import { AuthenticationByAccountRepository } from '@/use-case/session/protocols/authentication-repository'
+import { AuthenticationByAccountRepository } from '@/use-case/session/protocols/authentication-by-account-repository'
 import { SessionListByAuthenticationRepository } from '@/use-case/session/protocols/session-repository'
 import { DbSessionByAccount } from '@/use-case/session/session-by-account/db-session-by-account'
 import { makeAuthenticationByAccountStub, mockedAuthentication } from '../stubs/authentications'
